+++ R/random.R
@@ -1,23 +1,27 @@
+choose_seed <- function(supplied, cache){
+  previous <- get_previous_seed(cache = cache)
+  seed_conflict <-
+    !is.null(previous) &&
+    !is.null(supplied) &&
+    !identical(previous, supplied)
+  if (seed_conflict){
+    stop(
+      "You supplied a seed of ", supplied,
+      "to either make() or drake_config(). ",
+      "Your project already has a different seed: ", previous, ". ",
+      "Use read_drake_seed() to see for yourself.",
+      call. = FALSE
+    )
+  }
+  (previous %||% supplied) %||% 0
+}
+
 get_previous_seed <- function(cache){
   if (cache$exists(key = "seed", namespace = "config")){
     cache$get(key = "seed", namespace = "config")
   } else {
-    get_valid_seed()
+    NULL
   }
-}
-
-# From withr
-get_valid_seed <- function(seed = get_seed()){
-  if (is.null(seed)) {
-    runif(1L)
-    seed <- get_seed()
-  }
-  seed
-}
-
-# From withr
-get_seed <- function(){
-  get0(".Random.seed", globalenv(), mode = "integer")
 }
 
 # A numeric hash that could be used as a
@@ -27,4 +31,13 @@ seed_from_object <- function(x) {
   hash <- digest::digest(x, algo = "murmur32")
   hexval <- paste0("0x", hash)
   utils::type.convert(hexval) %% .Machine$integer.max
+}
+
+# From lintr
+`%||%` <- function(x, y) {
+  if (is.null(x) || length(x) <= 0) {
+    y
+  } else {
+    x
+  }
 }