+++ NEWS.md
@@ -1,5 +1,7 @@
 # Version 5.1.0
 
+- Add a `seed` argument to `make()`, `drake_config()`, and `load_basic_example()`. Also hard-code a default seed of `0`. That way, the pseudo-randomness in projects should be reproducible
+across R sessions.
 - Cache the pseudo-random seed at the time the project is created and use that seed to build targets until the cache is destroyed.
 - Add a new `drake_read_seed()` function to read the seed from the cache. Its examples illustrate what `drake` is doing to try to ensure reproducible random numbers.
 - Evaluate the quasiquotation operator `!!` for the `...` argument to `drake_plan()`. Suppress this behavior using `tidy_evaluation = FALSE` or by passing in commands passed through the `list` argument.